Move Deprecated LFSR16 to Compatibility
